refactor: groupForm 위치 변경
