+++ src/components/groupList/SubmitInviteForm.tsx
@@ -1,5 +1,5 @@
-import { useInviteGroupForm } from '@/hooks/byUse/useGroupForm';
 import { useInsertUserGroupMutation } from '@/hooks/queries/byUse/useGroupMutations';
+import { useInviteGroupForm } from '@/hooks/useCustomForm/useGroupForm';
 import { makeUserGroupDataToObj } from '@/services/groupServices';
 import { Button } from '@/stories/Button';
 import browserClient from '@/utils/supabase/client';
